Telling the customer why the should care about custom scripts
When I was helping a customer, I realised this is the landing page for custom scripts. Instead of telling the customer all the amazing super powers they get with scripts in Octopus, we told them about the security context and how to debug scripts in an antiquated way.

I've re-authored the page trying to put myself in the shoes of a customer arriving at Octopus for the first time.