Adds region failure to the tables
I was looking at this with @alexrolleyoctopus and @RayAtOctopus and found the region failure scenario was buried in a section at the end of the page. I expected to have a region failure estimate in the same tables as data center failure.

I've added as estimate to the table to more clearly show how the scenarios escalate, along with a pointer to the region failure section.